userGoal.dart
Got current time going for addition. I think where it is right now it is more of a total. I need to figure out how to reset all of the current times and add it to the total time at the end of the day.

Also how to save the current time for each day