Update cloudFirestore.dart
Trying to get the datesCompleted thing working.

I don't think that I really need the currenTotal in the data structure, I may be able to get rid of that.

Then I can use the map to keep track of currentTotals for each day and then total for well the total.