home.dart
Trying to get this favorite thing to work, right now it is very messy. Tomorrow I can try and finish these methods that I have going on right and now and then try and clean it up a bit regardless of the outcome.

Everything is feeling a little messy right now, but I think that I am getting pretty close to completing it.

I definitely need to do a full clean of everything, really I would need to restart to get everything perfect I believe. Consistent and clean. But I think it is within the margin of cleanness and I will accept it because I actually want to finish this.